init mapper

+++ api/src/main/kotlin/top/kanetah/planhv2/api/entity/SubmitFileAttributes.kt
@@ -6,6 +6,6 @@ package top.kanetah.planhv2.api.entity
 data class SubmitFileAttributes(
         val formerName: String,
         val saveName: String,
-        val size: Long,
+        val size: Double,
         val path: String
 )
